LayoutTests/fast/dom/HTMLInputElement/input-value-set-null-expected.txt

Checks that setting input.value to null sets the input value to the empty string.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


input.defaultValue = 'defaultValue'
PASS input.value is "defaultValue"
input.value = null
PASS input.value is ""
PASS successfullyParsed is true

TEST COMPLETE
